project structure refactor, contract files are only used for our tests
