+++ hardhat.config.ts
@@ -3,28 +3,28 @@ import "@nomiclabs/hardhat-ethers";
 import "@nomiclabs/hardhat-waffle";
 
 export default {
-  defaultNetwork: "mainnet",
+  defaultNetwork: "hardhat",
   networks: {
     mainnet: {
       url: process.env.MAINNET_PROVIDER_URL,
     },
     hardhat: {
-      // forking: {
-      //   url: process.env.FORKING_PROVIDER_URL,
-      //   blockNumber: 12464951,
-      // },
-      // mining: {
-      //   auto: true,
-      //   interval: 0,
-      // },
-      // allowUnlimitedContractSize: false,
+      forking: {
+        url: process.env.FORKING_PROVIDER_URL,
+        blockNumber: 12464951,
+      },
+      mining: {
+        auto: true,
+        interval: 0,
+      },
+      allowUnlimitedContractSize: false,
     },
   },
   paths: {
-    sources: "./src/contracts",
+    sources: "./test/contracts/src",
     tests: "./test/contracts",
-    cache: "./cache",
-    artifacts: "./artifacts",
+    cache: "./test/cache",
+    artifacts: "./test/artifacts",
   },
   solidity: {
     version: "0.7.6",
@@ -42,7 +42,7 @@ export default {
     timeout: 3600000,
   },
   typechain: {
-    outDir: "src/typechain",
+    outDir: "test/typechain",
     target: "ethers-v5",
     alwaysGenerateOverloads: false,
   },
