+++ test/contracts/src/contracts/libraries/LICENSE_GPL
